Major revisal -- several bugfixes --
* fixed bug with accessing Piece-Square-Table
* (feature) new alpha values are always searched -> no more PV shortening -> major speed boost in iterative deepening

+++ ChessEngine/hash_entry.h
@@ -3,8 +3,8 @@
 class hash_entry
 {
 public:
-	hash_entry(double set_score, int set_depth_left);
-	double score;
+	hash_entry(int set_score, int set_depth_left);
+	int score;
 	int depth_left;
 	unsigned times_hit;
 	~hash_entry();
